requirements.txt
# LangChain and core dependencies
langchain>=0.1.0
langchain-community
langchain-google-genai
langchain-pinecone

# Google Generative AI
google-generativeai

# Embeddings and Transformers
sentence-transformers>=2.2.2
huggingface_hub>=0.14.1

# Pinecone vector store
pinecone-client
pinecone

# PDF parsing
pypdf

# Flask for API server (optional)
flask

# Jupyter notebook (optional, for .ipynb support)
notebook
ipykernel

# Environment handling (recommended for .env files if needed)
python-dotenv

sentence-transformers         # Text embedding model (MiniLM)
huggingface_hub         # Compatible version for sentence-transformers

-e . # Local package for the project
